chore: move field-types to base-application/internal/types
