+++ generators/base-application/entity-all.d.ts
@@ -22,7 +22,7 @@ import type { Entity as BaseEntity } from '../../lib/types/base/entity.js';
 import type { ServerEntity } from '../server/types.js';
 import type { Field as BaseField } from '../../lib/types/base/field.js';
 import type { Relationship as BaseRelationship } from '../../lib/types/base/relationship.js';
-import type { FieldType } from '../../lib/application/field-types.ts';
+import type { FieldType } from './internal/types/field-types.ts';
 import type { FakerWithRandexp } from './support/faker.ts';
 import type { Field } from './field-all.js';
 import type { Relationship } from './relationship-all.js';
