+++ lib/jdl/converters/jdl-to-json/jdl-to-json-field-converter.ts
@@ -25,7 +25,7 @@ import type JDLObject from '../../core/models/jdl-object.js';
 import type { JSONField } from '../../core/types/json-config.js';
 import type { JDLEntity } from '../../core/models/index.js';
 import type JDLField from '../../core/models/jdl-field.js';
-import type { FieldType } from '../../../application/field-types.js';
+import type { FieldType } from '../../../../generators/base-application/internal/types/field-types.ts';
 
 const {
   Validations: { UNIQUE, REQUIRED },
